Shutdown cleanly when receiving a Windows console shutdown request.
This commit makes tinc exit cleanly on Windows when hitting CTRL+C at
the console or when the user logs off. This change has no effect when
running tinc as a service.
